[ADD] aligned spacing of main content w/ header on mobile, translation adjustments, fixed model import, renamed post-overview.service to post.service
