+++ src/app/features/header/components/blog-overview/blog-overview.component.scss
@@ -6,7 +6,6 @@
 
 .blog-overview-field {
     font-size: $large-font-size;
-    margin-left: 1em;
     width: 50%;
 
     @media only screen and (max-width: $md) {
