+++ src/app/features/header/container/header.component.scss
@@ -14,10 +14,6 @@
     .nav-right {
         display: flex;
         align-items: center;
-        
-        .logout-btn {
-            margin-right: 1em;
-        }
     }
 }
