+++ src/app/features/login/container/login.component.html
@@ -10,8 +10,8 @@
   </div>
   <mat-card class="login_card">
     <img alt="Login to Blogger" src="https://img.icons8.com/fluent/96/000000/lock-2.png" />
-    <h1>Welcome to Blogger!</h1>
-    <p>Log in with a Google account to start using Blogger.</p>
-    <button mat-flat-button color="accent" (click)="signIn()">Sign in</button>
+    <h1>Willkommen zu Blogger!</h1>
+    <p>Log dich mit deinem Google Account ein, um fortzufahren.</p>
+    <button mat-flat-button color="accent" (click)="signIn()">Einloggen</button>
   </mat-card>
 </div>
