+++ src/app/features/post-overview/container/post-overview.component.scss
@@ -32,8 +32,14 @@
     right: 0;
     margin-bottom: 4em;
     margin-right: 4em;
+    z-index: 100;
 
     display: flex;
     justify-content: center;
     align-items: center;
+
+    @media only screen and (max-width: $md) {
+        margin-bottom: 1em;
+        margin-right: 1em;
+    }
 }
